requirements.txt
# Core dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
python-dotenv==1.0.1

# Model and ML dependencies
torch>=2.0.0
torchvision>=0.15.0
# Note: Install diffusers from git for latest Qwen support:
# pip install git+https://github.com/huggingface/diffusers
diffusers>=0.31.0
transformers>=4.40.0
accelerate>=0.25.0

# Image processing
Pillow>=10.0.0

# Data validation and serialization
pydantic==2.10.3
pydantic-settings==2.6.1